[wasm][type-reflection] Fix nested WasmJSFunction
Follow-up of:
6225722: [wasm][type-reflection] Do not unwrap WasmJSFunctions on import
https://chromium-review.googlesource.com/c/v8/v8/+/6225722

The previous change would still return
ImportCallKind::kJSFunctionArityMatch/Mismatch but link the outer
WasmJSFunction as the target, which is incorrect. WasmJSFunctions should
be called with kUseCallBuiltin wrappers instead.
Revert to the old behavior of linking the inner callable for the
standard case, and treat the nested case separately with the appropriate
import kind.

R=jkummerow@chromium.org

Fixed: 394767640
Change-Id: If97cf9021d572e512f9be10c57767f9078ffca11
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6237577
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98582}