[float16array] Fix fp64->fp16 representation change again
The current way of eagerly inserting a NumberToFloat16RawBits node is
incorrect, because it is typed as a Number. This means it can be
constant folded away with a number constant, which will be the incorrect
representation.

This CL removes the higher level NumberToFloat16RawBits operator and
introduces a MachineRepresentation::kFloat16RawBits for the purposes of
representation conversion. As output, it is eagerly changed to a
Float64. As input, it is first converted to float64, then truncated to
float16, and finally bitcast to a word16.

kFloat16RawBits is never used after machine lowering (similar to
kMapWord being unused after memory lowering). Machine lowering always
lowers uses of kFloat16RawBits to uint16.

Fixed: 394327270
Bug: 42203953
Change-Id: I2e3fbc9d5956abbfc89202c004457483f2b51be8
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6245152
Commit-Queue: Shu-yu Guo <syg@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98613}