[js_fuzzer] Provide stub for AsyncIterator
Some external test suites like Spidermonkey contain tests for:
https://github.com/tc39/proposal-async-iterator-helpers

Having the AsyncIterator as a global is not implemented in V8
and requires an additional stub.

This adds logic to the identifier normalizer pass to also check
for identifiers that require additional resources.

This loads a stub for AsyncIterator only if the respective
identifier is used in the code.

Bug: 389069288
Change-Id: I019a8e851479ee344b4b73329d860ef3b2b31542
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6244288
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Stephen Röttger <sroettger@google.com>
Cr-Commit-Position: refs/heads/main@{#98604}