[roots] Expose single-char strings as roots, use in parser
Create root indices for all 256 single-character strings, so that they
can be accessed by name (not just the single character table), and the
root can be used as a handle. A python script is used to generate the
macro list for these, along with friendly names suggested by a friendly
AI.

Use this new functionality to initialize the Isolate-wide
AstStringConstants with the ASCII subset of one byte characters, so that
we don't need to initialize (and hash!) them lazily on each parser
invocation.

Bug: 395570874
Change-Id: I7ee9c41f61cf94068447757b80496e725b28a681
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6249318
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98612}