[js_fuzzer] Wrap var with try-catch
If a variable declarator uses a destructuring expression, we can't
catch errors if the try-catch wrapper function returns undefined.
E.g.:
let {a, b} = __wrapTC(() => ...);

While the executed code in the initializer won't throw an exception,
it might not execute well and cause an exception when destructuring.

We can't really solve this for `const` and `let`, but for `var` we
now keep the old try-catch wrapping form, which covers also the
destructuring.

Bug: 389069288
Change-Id: I22a59b73e33786f4d3fbb8a1ed584cfbb549cbd1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6249323
Reviewed-by: Stephen Röttger <sroettger@google.com>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98628}