[string] Fix overflow case of TryParseArrayIndex
The "undo" on overflow was undoing the overflowing add, but not the
overflowing multiplication, which is lossy and not undoable. This is
probably being too clever -- on 64-bit platforms, where this overflow
can happen, we can just have a 64-bit index variable and arithmetic on
it will still be cheap. On 32-bit platforms, we can keep the clever
check in order to keep a 32-bit value (and cheap 32-bit arithmetic);
these don't need an overflow case here anyway.

Drive-by fix a lint error with anonymous namespaces in a header.

Fixed: 395053819
Change-Id: I77d79afdaaef371c4b042ae3dffba69b19f9684b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6249327
Commit-Queue: Patrick Thier <pthier@chromium.org>
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98624}