[turbofan] Use AdditiveSafeInt feedback for faster int add/sub
Add Int53 addition feedback and use it during simplified lowering
to optimize integer addition and subtraction.  When one of the inputs
is known to be within the Int53 range (since the check is relatively
expensive), we can use a shifted integer addition instead of floating
addition.

AdditiveSafeInteger (Int53) is used instead of SafeInteger because
the safe integer range (-2^53 + 1 to 2^53 -1) excludes -2^53.
Using SafeInteger directly would require an additional check after
the operation to ensure the result remains within the valid bounds.

This change does *not* implement Maglev-specific optimizations.

Bug: 384959125

Change-Id: I7b54d075c55ab7828e407cec03933a34be6247d5
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6038008
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98643}