[compiler-dispatcher] Reuse task pool across isolates
This CL moves the OptimizingCompileTaskExecutor into the IsolateGroup
to share the same input queue and thread pool across multiple
isolates.

Isolates do not need to create their own instance anymore and just
use the one from the IsolateGroup.

Note that we currently do not destroy the default IsolateGroup
in e.g. V8::Dispose. Which means that we don't really run the
destructor for OptimizingCompileTaskExecutor anymore with this CL.

This seems to improve bomb-workers by about 3.7% according to
pinpoint.

Bug: 391169065
Change-Id: Ia4d9bbd83c83e0ffe5376f0f64da8a3346f98a99
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6250138
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98640}