[wasm][fuzzer] Return -1 for non-interesting inputs
This guides corpus generation in libfuzzer. Return -1 for any inputs
which should not be added to the corpus.
We also return -1 for cases which are still somewhat interesting (any
invalid modules also has a chance to hit a decoder bug). We don't want
those cases in the corpus though, as it's hard enough to generate valid
modules for fuzzers. We will get enough invalid modules via random
mutations.

R=machenbach@chromium.org, mliedtke@chromium.org

Bug: 384781857
Change-Id: I7c8833c45cb1dc8151bacb130906352c65f7511d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6180311
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98654}