[js_fuzzer] Avoid collision with secondary module builder
The wasm-module-builder lives in V8 and in an older version in the
Chakra test suite. Adding both at the same time leads to bail-outs
due to duplicate variable declarations.

This hard-codes the removal of the older version in case both are
present. This code could be generalized in the future for other
collisions if there are any.

Patch 2 adds a regression test with the scenario that currently
fails. The test actually executes the generated test case with node.
https://crrev.com/c/6253701/2

Patch 4 adds the change to drop one of the module builder
implementations.
https://crrev.com/c/6253701/2..4

Bug: 389069288
Change-Id: I6a6b73fa8361585de56ab53d9339d7f50607db9a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6253701
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Stephen Röttger <sroettger@google.com>
Cr-Commit-Position: refs/heads/main@{#98652}