[heap] Let RegionAllocator use the last segment of a region
In case a virtual address space is used to allocate segments that can
completely fill the space, then the region allocator will fail to
allocate the last segment. The reason is that it always adds the
padding, even if it might not be needed, since the base itself has the
required alignment.

Bug: 395536898
Change-Id: Ic212fa9277f32c55f54954894bffe08ee9e87f43
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6257548
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Commit-Queue: Olivier Flückiger <olivf@chromium.org>
Auto-Submit: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98682}