[interpreter] Check if we want to tier up before OSRing to baseline
We deopt to the interpreter instead of baseline because of CFI. If we
have OSR'd code that deopts after the loop because of missing
feedback, we'll be in the interpreter when we JumpLoop on an outer loop. With the previous mechanism we'd tier up to baseline on
that JumpLoop, and OSR to the cached optimized code on the inner loop; after which we'd deopt again at the end of the inner loop. We'll never
manage to actually optimize the outer loop; and we also don't generate
better OSR code for the inner loop

By letting the interpreter check its optimization status on OSR points
we have a chance of getting better code at the outer loop as well.

Change-Id: I0455797cbb2935ab49a21751d6c809ae85e76ccf
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6263711
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Commit-Queue: Olivier Flückiger <olivf@chromium.org>
Auto-Submit: Toon Verwaest <verwaest@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98680}