Fix scope/context mismatch when pausing in for-of loops
This CL fixes a long-standing issue where the debugger can crash when
we pause in the right-hand side expression of a for-of loop header.

The reason is that source-position wise, the expression is part of
the parser inserted block scope that make let/const/var declarations
in the left-hand side work. This inserted block-scope may require
a context (e.g. when the let/const/var declared loop variable is
captured in a closure).

When the debugger pauses in the right-hand side expression, then
it pauses in this inserted block scope, but the debugger can't find
the block context as the bytecode didn't create and push it yet.

There are 2 potential solutions to fix this:

  1) Change AST scopes so they can consist of two source ranges,
     not just one. This would allow the parser to exclude the right-
     hand side expression from the inserted block scope while
     including the left-hand side.

  2) Put the right-hand side expression in a separate scope and
     make it a sibling of the other inserted block scope. V8
     scopes don't have to be nested properly w.r.t. to source
     positions, so this also works.

This CL implements option 2).

R=leszeks@chromium.org

Fixed: 41338202,41349575,383029265
Change-Id: Ic71d702b207fda7e4654969a55997cfc7dfb606e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6268480
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98700}