[tiering] Mark optimize to TF on M->TF OSR
If the TF OSR request came in while the function was still marked for
optimization to M, we'll ignore the request to tier the function up to
TF at that point. This is (possibly) useful since maglev compiles
quicker than TF.

The problem is though that TF OSR code doesn't manage interrupt
budgets, so if we always enter a function through M code (once ready)
and then OSR to TF, we'll never change the optimize the function
through the entry-point.

This change hooks into OSR from M -> TF to mark the function for
tierup to TF if we already managed to install M code in the meantime.

Change-Id: I8fc2c03471bc28e59b1409631edd8a943cf7c7d8
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6268486
Commit-Queue: Toon Verwaest <verwaest@chromium.org>
Auto-Submit: Toon Verwaest <verwaest@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98702}