[json] Stringify: Avoid checks on object/array resume
We resume array/object serialization every time a property/element is
an array/object.
I.e. for [[],[],[],...] we resume after every element.
Previously we just resumed by setting the start index, but still did
several checks on the object to ensure we can serialize it on the
fast-path. This is unnecessary as we already checked the object in the
beginning.
We therefore split off iterating the elements/properties and avoid those
checks when we resume.
This requires a distinction in ContinuationRecord types whether we
resume or serialize an object from the beginning, plus more types for
array resumption to resume on the correct path without needing to check
for element kinds again.

Bug: 380044242
Change-Id: I3546b085c2b22679d0eae505b9ef37b9947117e2
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6233897
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98754}