[wasm] Do not get the isolate from heap objects
Under the sandbox attacker model this is inherently unsafe: As the
(in-sandbox) heap can be manipulated arbitrarily, we could read an
object belonging to another isolate. In most cases where we call
`JSReceiver::GetIsolate()` we then assume that this is the "current
isolate" and access this isolate's trusted pointer table, or allocate
handles in that isolate.

The fix is to use thread-local storage and read the current isolate from
`Isolate::Current()`.

R=jkummerow@chromium.org

Bug: 396607238, 393402168
Change-Id: I76d6dedd507559814516fb6b0898bdb95ea1928e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6269818
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98730}