[js_fuzzer] Determine loop variables during context analysis
For subsequent changes to decrease infinite loops, we need to
know which variables are referred to in loop conditions.

Babel's traversal, however, doesn't easily allow to differentiate
between the test and body parts of a loop. We therefore add a stack
based analysis to the context analyzer that keeps track of loop and
block statements. All variables we discover while in a loop, but
not in a block statement, are considered loop variables.

The normalizer now turns all block-less loops into loops with block
statements as bodies, which should not reduce V8 fuzz coverage as
this only affects the parser.

This adds a test that demonstrates the determined loop variables and
shows normalized sources. Further use of the information will be
added in subsequent CLs.

Bug: 389069288
Change-Id: I120f56c3cfa062c606dfe40eede0206317cc22d0
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6275924
Reviewed-by: Carl Smith <cffsmith@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98739}