Reland "[compiler-dispatcher] Reuse task pool across isolates"
This is a reland of commit c3da216513c16cd14bb86a0c7bc0939f9eea2c84

The reason for the revert was that so far the thread pool wasn't
cancelled. This can become an issue in d8 when we destroy the default
platform. At this point compiler tasks may be still running.
In the flaky crash in https://crbug.com/396091278 the compiler threads tried to emit a trace event and needed to access the platform to do so.

This issues should have been fixed by https://crrev.com/c/6276488.
The linked CL destroys the default IsolateGroup in V8::Dispose and
with it the thread pool. This allows us to reland this CL unmodified.

Original change's description:
> [compiler-dispatcher] Reuse task pool across isolates
>
> This CL moves the OptimizingCompileTaskExecutor into the IsolateGroup
> to share the same input queue and thread pool across multiple
> isolates.
>
> Isolates do not need to create their own instance anymore and just
> use the one from the IsolateGroup.
>
> Note that we currently do not destroy the default IsolateGroup
> in e.g. V8::Dispose. Which means that we don't really run the
> destructor for OptimizingCompileTaskExecutor anymore with this CL.
>
> This seems to improve bomb-workers by about 3.7% according to
> pinpoint.
>
> Bug: 391169065
> Change-Id: Ia4d9bbd83c83e0ffe5376f0f64da8a3346f98a99
> Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6250138
> Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
> Reviewed-by: Leszek Swirski <leszeks@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#98640}

Bug: 391169065
Change-Id: I4d0aa2b0112207fe7be3c0a51fbb165203a013db
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6268679
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98790}