[js_fuzzer] Limit fuzz-library calls
Calling random functions or using random objects or properties gets
very inefficient when it happens in large loops.

This adds a safeguard limit that prevents further execution of
such functions if an overall call amount is exceeded. In particular
random function calls lead to performance issues, so we limit those
to 25, while random property or object accesses are limited to 50.

Bug: 389069288
Change-Id: I7d41f3185ead044884f3e103e6915c4a9a7811f5
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6276577
Reviewed-by: Carl Smith <cffsmith@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98762}