[sandbox] Introduce GetCurrentIsolateForSandbox
This is the safer alternative to GetIsolateForSandbox which reads the
isolate from an on-heap object and can thus return an isolate which is
not the "current" isolate.

As we now include "src/execution/isolate-inl.h" in
"src/sandbox/isolate-inl.h" we also need to resolve a few missing
includes or missing declarations.

R=jkummerow@chromium.org

Bug: 396607238, 393402168
Change-Id: I697bb22a92b625868dc69b9054a14218e2002c47
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6276454
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98791}