[ia32][regalloc] Only use register hint if it is allocatable
On ia32 with AVX disabled, pblendvb has xmm0 as a fixed register input,
and the register allocator uses that as a hint to assign xmm0 to the
input live range. However when it applies the hint, it never
checks that it belongs to the allocatable subset of registers. And ia32
is an exception: xmm0 is not allocatable, so using the hint is invalid
in this case.

R=dlehmann@chromium.org
CC=​clemensb@chromium.org

Fixed: 396460489
Change-Id: I1df091400e1b75b4b7d64f310383ce58a1bc7386
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6277587
Reviewed-by: Daniel Lehmann <dlehmann@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98763}