[js_fuzzer] Prevent block-wrapping nested function declarations
A variable and function declaration with the same name might live
side-by-side in a function body, but if they move up by one block,
some scope compatibility rules make it a syntax error:
https://tc39.es/ecma262/#sec-block-level-function-declarations-web-legacy-compatibility-semantics

A recently introduced try-catch behavior sometimes tries to try-catch
wrap an entire function body, which can lead to problems if there's
a nested declaration in the body. We fix this now by just falling back
to the old behavior in this case (single statement wrapping).
Actually checking for duplicate identifiers in the fuzz AST is
probably overkill.

Patch 1 adds a regression test to show the status quo:
https://crrev.com/c/6277585/1

Patch 2 adds the fix:
https://crrev.com/c/6277585/1..2

Bug: 389069288
Change-Id: Ifc7a215cf040af3a67db1d9a448fb0943afb62c4
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6277585
Reviewed-by: Carl Smith <cffsmith@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98770}