[fastapi] Check array length before cast to FixedDoubleArray
The backing store of a JSArray whose element kind is
`PACKED_DOUBLE_ELEMENTS` is typically a `FixedDoubleArray`, except if
the array is empty, in which case it's the EmptyFixedArray. Fast api
code, however casted to `FixedDoubleArray` unconditionally, which
triggered a DCHECK when the the backing store was an EmptyFixedArray.

Bug: 395028748
Change-Id: I5cac7cdb82986d62fbe0fbae0f9665d83e2c159c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6280805
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98792}