[wasm] Add sample for '0 deopts' for potentially deopting functions
Note that this sample is only created when compiling a function with
the optimizing compiler if it has any deopt data.
So, if the optimized code is served from cache and triggers a deopt
later, there will be a "1 deopt" sample but no corresponding
"0 deopts" sample. While this means that numbers are a bit imprecise,
it is probably preferred over adding "0 deopts" samples for each
function served from cache with a deopt point (as looking it up from
cache happens somewhat eagerly and the function might not be run at
all.)

Bug: 42204618
Change-Id: I2813226d7a69758420b66cfdfb2be0124231db7b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6281086
Auto-Submit: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Daniel Lehmann <dlehmann@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98793}