[test] Enable testing the jit-fuzzing variant
Run Mjsunit with --jit-fuzzing. To prevent any hassle with
flag contradictions, we also use --fuzzing. Additionally this
requires --no-fail, as --jit-fuzzing breaks many test assumptions.

This would have caught http://b/395032421 deterministically.
An example from the temporarily enabled trybot is:
http://ci.chromium.org/b/8722502487644513649

For now, we enable this only on the Linux64 debug CI bot to get a
signal if and when it fails. This closes the tree and blocks LKGR,
which would prevent rolling a V8 version with such a breakage in
the future.

Bug: 395032421
Change-Id: Idaef8fa9ac1cb3978e08821dd94c40977da96b43
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6277292
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98819}