Prevent flakes on sandbox tester to close the tree
The sandbox tester is the only builder running tests and closing the
tree, while situated on the clusterfuzz console, which uses a
different builder abstraction v8_builder compared to other
tree-closing builders.

The logic to determine the proper v8 notifier to ignore flakes
when closing the tree, was missing there. This closed the tree
wrongly a couple of times on flakes.

Bug: 365967057
Change-Id: I2a9a67a90fe9c189999b90b155b74a2f1a558087
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6285845
Auto-Submit: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>