[sandbox] Remove invalid DCHECK in CppHeapPointerTableEntry::Mark()
This DCHECK does not hold after all. On the mutator we perform
basically three operations: 1) Initialize 2) Update Pointer 3) Mark.

When the concurrent marker now observes the value of 1) and tries to
set the mark bit between 2) and 3), the concurrent marker will see
the updated value from 2) but not the markbit yet.

This is because we do not set the new value and the markbit
atomically in JSApiWrapper::SetCppHeapWrappable.

Bug: 396621723
Change-Id: I66a1158661446775802aec91f8cb8482bbf687c9
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6285885
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98817}