[js_fuzzer] Fix unavailable function in old Node version
The fuzzer is built with `pkg` and the latest available Node version
with this tool is Node 18. That version doesn't include recent
Javascript Set functions yet.

The unittests unfortunately run with Node 23 and didn't catch this.
Will look into this in a follow up to see if we should downgrade
those.

Bug: 389069288
Change-Id: I67fdcfffe4c681ff656b69a1584ea78dfdf73c48
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6286469
Auto-Submit: Michael Achenbach <machenbach@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Stephen Röttger <sroettger@google.com>
Cr-Commit-Position: refs/heads/main@{#98852}