[compiler] Avoid lock on allocation in JSNativeContextSpecialization
JSNativeContextSpecialization::Concatenate was allocating the new
string while holding the lock acquired by
SharedStringAccessGuardIfNeeded. This mutex was recently changed to
an exclusive lock, which can cause deadlocks in certain scenarios.

For example the main thread could try to acquire the lock and block
because the background compilation already holds the lock in
Concatenate. The compilation thread then allocates the result
string object. If this allocation triggers a GC, we end up in a
deadlock. This is because a GC needs to happen on the main thread but
the main thread is already blocked on a mutex. And the background
thread will not unlock the mutex because it needs the GC to happen
first.

This CL fixes this by locking the mutex only after the allocation.

Bug: 398064853
Change-Id: Iaae504f15d4cc756bb64b9bd2947aa6ca2598de1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6287330
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Patrick Thier <pthier@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98865}