[sandbox] Check that GetIsolateForSandbox returns the current sandbox
We want to replace all calls to GetIsolateForSandbox via
GetCurrentIsolateForSandbox. Before doing that, this CL adds a SBXCHECK
to verify that both methods return the same isolate.

One exception are shared objects, where we do return a pointer to the
main isolate instead. For those cases we do check that the isolates are
"compatible" in the sense that they share the same shared pointer table.
This is a pretty costly check, but it is not executed in production code
and it should soon (maybe after the next branch cut) be removed by
replacing all callers of `GetIsolateForSandbox` by
`GetCurrentIsolateForSandbox`.

R=dinfuehr@chromium.org
CC=mlippautz@chromium.org

Bug: 396607238, 393402168
Change-Id: I3a5bc983b7e79db157f8dfe5088a327fc2a75e32
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6276459
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98872}