[heap] Parallelize processing of young wrapper reclamation
Weak young wrapper objects can be reclaimed (i.e., treated weakly)
when they are otherwise unreachable and follow certain conditions.
Processing is split into 3 phases:
1.  computing weakness;
2.  processing strong objects;
3.  clearing weak objects;

This CL implements parallel processing for phases 1. and 3. It revives
an idea that was first explored in issue 40264139.

Since the initial inception the area around traced handles has been
reworked and so the implementation can now nicely parallelize based on
blocks (TracedNodeBlock) instead of creating an artificial view on top
of a large vector.

The main idea of the new approach is:
A.  Free the handle concurrently on a block which is exclusively owned;
B.  Post process block afterwards which should be cheap;

Parallel processing is currently disabled behind
  --js-flags=--parallel_reclaim_unmodified_wrappers

Impact on S3:
👍 TodoMVC-jQuery                              [ -1.3%,  -0.1%]

Bug: 398773905
Change-Id: I8354309381f58fcc01408ab489f7060f627f9166
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6286229
Reviewed-by: Anton Bikineev <bikineev@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98887}