[heap] Add Heap::StartResizeNewSpace() for MinorMS
MinorMS resizes the new space in two steps. With this CL MinorMS can
now use the StartResizeNewSpace() and ResizeNewSpace() methods on the
Heap class to do so. This change allows us to share the computation
for the reduced new space size more naturally.

Bug: 391129951
Change-Id: Idd566157d93c1bebf26483675c9907a833aecbea
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6298738
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98912}