[compiler] Collect constness dependencies OffTheRecord
This is a relands half of
commit 0a5f21efef28310bd8b5311481699221e0d8c0a2

Skipping constness dependencies in Maglev was incorrect as there
actually is a downstream optimization. In particular we are not
updating the constness property in maglev stores and instead bail out
if the access info indicates that the field should be constant.

Change-Id: Ic81c1fdd2bc1f4b1ff34deb1c470e275a1e7e22c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6297951
Auto-Submit: Olivier Flückiger <olivf@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98909}