[wasm][jspi] Clear wasm stack EPT entry on exit
Null the stack pointer in the EPT when the stack retires. If a
continuation object was corrupted inside the sandbox to keep a reference
to this EPT entry, this ensures that this object cannot be used to cause
a UAF of the stack, but will result in a null-dereference instead.

R=mlippautz@chromium.org,cffsmith@chromium.org

Bug: 384549659
Change-Id: I72f23db82955c3e736ab3017ae1fb221e7533f4f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6264503
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98972}