[js_fuzzer] Improve try-catch wrapping with structured returns
Sometimes, functions return object expressions which are
destructured on the call site. When the entire function body is
wrapped with try-catch, an empty catch results in an undefined
return value, which causes a type error when being destructured.
As the destructured initializers are often declared with let or
const, they are also not try-catch wrapped, which leads to
bailouts. This pattern occurs quite often in initialized wasm
modules.

E.g.
function foo(){ return {a: 0, b: 1}; }
let {a: a, b: b} = foo();

We now also return an empty object expression in such cases,
which is enough to prevent destructuring errors. With an
empty object all destructured values are undefined.

Patch 1 extends an existing test:
https://crrev.com/c/6298126/1

Patch 2 shows the new logic:
https://crrev.com/c/6298126/1..2

Bug: 389069288
Change-Id: I181bc99737bfc8a6a0760ca62bb643635527e16a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6298126
Reviewed-by: Carl Smith <cffsmith@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98995}