[maglev] Fix string length on non-string string wrapper inputs
In dead code it can happen that a string wrapper input is not a string
which ends us in a DCHECK if we are not careful.

The reason this can happen is that the input to UnwrapThinString might
be typed string, but its tagged alternative not.

Fixed: 399208503
Change-Id: I88b6215ed4f4ff98861e52ed8a01df95c494571c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6304622
Auto-Submit: Olivier Flückiger <olivf@chromium.org>
Reviewed-by: Patrick Thier <pthier@chromium.org>
Commit-Queue: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98941}