Introduce V8Inspector::consoleTimeStampWithArgs
Which accepts the arguments passed to console.timeStamp. On the Blink side implementation [1], it will handle the arguments and convert performance.now() timestamps to the tracing clock when needed (as per go/cpq:console-timestamp-numbers).

Once the Blink side is implemented we can remove the implementation in V8 and forward the call to Blink via V8Inspector::consoleTimeStampWithArgs [2].


[1] https://chromium-review.googlesource.com/c/chromium/src/+/6311629
[2] https://chromium-review.googlesource.com/c/v8/v8/+/6311610

Bug: 399851200
Change-Id: Idfce7f59299a151bf452ad6351325c6812bbcff1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6311609
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Eric Leese <leese@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99028}