[turbofan] Ensure safe add int truncation check
Ensure we check if the type_check is additive safe integer when
truncating to word32 first, since we are speculating that the input
is in the safe additive integer range.

Previously, if the output was a number, then the simplified lowering
had checked statically that the operation were in the safe
integer range. So no runtime checks were needed.

Fixed: 399698534
Change-Id: Id91c99e2006b80360bb616c28efc2ccde52f7872
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6309564
Commit-Queue: Nico Hartmann <nicohartmann@chromium.org>
Auto-Submit: Victor Gomes <victorgomes@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99022}