[wasm] Support --no-wasm-opt in Turboshaft (in DEBUG config)
This flag wasn't supported in Turboshaft as the
MemoryOptimizationReducer is run as part of the WasmOptimizePhase
which is skipped with --no-wasm-opt.

There are multiple bad options:
1) Split it into a separate phase causing potentially compile time and
   binary size regressions.
2) Add a separate phase that is only used for --no-wasm-opt. This will
   still significantly increase binary size.
3) Kill the --no-wasm-opt flag. Nobody seems to have missed it since
   we switched over to Turboshaft.

While I'm leaning towards (3), this CL opts for (2) in case anyone
would like to to skip optimization phases in Turboshaft completely.
(We already have more specific flags like
--no-turboshaft-wasm-loop-unrolling, so I'm not sure how much desire
there is for that.)
To prevent binary size regressions, this phase only gets compiled if
the code is compiled with the DEBUG macro (similar to e.g. some
tracing flags).

Fixed: 399412227
Change-Id: Icc998ac856e17ffdb67409e4c60864f7ddb3597e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6316667
Auto-Submit: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99027}