[array] Improve JSArray capacity changes
Change the Array length setter to allocate a FixedArray with exactly the
right length, rather than bumping up the old array capacity -- this
prevents the strangeness that

    a = []; a.length = kMaxLength

is ok, but

    a = []; a.length = kMaxLength - 1; a.length = kMaxLength

throws a RangeError (or crashes). We anyway expect that array length is
set manually in order to pre-size the array to the right length, and
this brings shortening and growing via length closer together.
Additionally, document this behaviour as well as documenting the fact
that setting array length will result in a dense FixedArray backing
store even if it is (initially) initialized with just holes.

Similarly, amend NewElementsCapacity to limit itself to
kMaxFixedArrayCapacity, so that arrays growing on push close to the max
size limit still have a chance to work.

Finally, change the kMaxFixedArrayCapacity to a power of two, because
powers of two are used elsewhere for capacity limits.

Change-Id: I7aec844e699bb0537905a402bb1cfe243c1db71f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6316907
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99031}