[wasm] Freeze registers on AddOutOfLineTrap and adapt first usages
This introduces a new interface for AddOutOfLineTrap() which freezes
the registers, so that the safe point attached to the out of line trap
location in debugging mode cannot be invalidated.

Most usages (excluding all the cases of division / modulo) are adapted
as part of this change.

Bug: 394238792
Fixed: 400093649
Change-Id: Ib269b75c30a478c393360914bedb9db32242a81a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6250139
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99039}