[runtime] Refactor TemplateInfo instantiation cache
This is a preliminary step for making it possible to use TemplateInfo
objects as keys in EphemeronHashTable. This CL ensures that once
serial_number is assigned to a TemplateInfo object it never changes
and thus it can be used for calculating TemplateInfo's hash.

Drive-by: split TemplateInfo into TemplateInfo (stuff related to caching
machinery) and TemplateInfoWithProperties.

Bug: 399393891
Change-Id: I24402d3d693cb74ce7d360452a1c88fb34014412
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6328038
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99072}