[wasm][jspi] Validate the in-sandbox chain of stacks
Validate that the in-sandbox chain of WasmContinuationObjects has not
been corrupted. Do this by also building a linked list of JumpBuffer
objects in the C++ heap when we resume a stack, and compare the pointers
when we return or suspend from a stack.

Update and validate the JumpBuffer in the C call that is performed on
every switch, and rename the C function from "SyncStackLimit" to
"SwitchStacks" to reflect its more general purpose (it already had more
responsibilities than just updating the stack limit).

Updating and validating the pointers requires that we pass the old stack
to the C function, which was not needed so far. Make the necessary
changes in the assembly builtins to keep the register alive and pass it
to the C call.

R=jkummerow@chromium.org

Bug: 384553540
Change-Id: If588c514b3a5f65ee134b0f591942bd2a6210126
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6328642
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99095}